rag_qa/requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
ipykernel
streamlit
langchain_groq
pypdf
langchain-ollama
faiss-cpu
